Move support classes to their own files
