+++ tests/Support/IntelDockerTags.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Tests\Support;
+
+use App\Shell\DockerTags;
+
+class IntelDockerTags extends DockerTags
+{
+    protected function platform(): string
+    {
+        return 'x86_64';
+    }
+}
